Add files via upload
Today, a lot of businesses operate online, making internet marketing crucial for retaining clients. However, it is not very effective to treat all customers equally and use the same marketing tactics to target them all. Instead, it irritates the customers by disregarding their uniqueness, hence customer segmentation is quickly gaining popularity and emerging as an effective remedy for this issue. Customer segmentation is the process of categorising a company's clientele based on behavioural (such as the kinds of items they order and their yearly income) and demographic (such as age, gender, and marital status) characteristics. Behavioural factors are a superior technique for customer segmentation since demographic indicators do not place an emphasis on the individuality of the consumer because same-age groups may have distinct interests.